+++ css/main.css
@@ -1,0 +1,9 @@
+.test-color {
+background-color: #fc0000;
+}
+
+.drop-area{
+    width: 10rem;
+    height: 10rem;
+    background-color: #9ec4a8f5;
+}
