+++ index.html
@@ -1,0 +1,70 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="UTF-8">
+    <title>ひづけるapp</title>
+    <link rel="stylesheet" href="node_modules/bootstrap/dist/css/bootstrap.min.css">
+    <link rel="stylesheet" href="css/main.css">
+    <meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline';" />
+</head>
+<body style="background: white;">
+    <h1>Hello World!</h1>
+    <p>
+        We are using Node.js <span id="node-version"></span>,
+        Chromium <span id="chrome-version"></span>,
+        and Electron <span id="electron-version"></span>.
+    </p>
+  <div id="drop-area" class="drop-area">
+    ここにドラッグ&ドロップできる
+  </div>
+    <button id="button1">ボタン</button>
+<script src="js/renderer.js"></script>
+
+  <script>
+    var holder = document.getElementById('drop-area');
+    /** hoverエリアにドロップされた */
+    holder.ondrop = function (e) {
+      e.preventDefault(); // イベントの伝搬を止めて、アプリケーションのHTMLとファイルが差し替わらないようにする
+
+      var file = e.dataTransfer.files;
+        
+        //console.dir(e);;
+
+      return false;
+    };
+async function scanFiles(entry, tmpObject) {
+    switch (true) {
+        case (entry.isDirectory) :
+            const entryReader = entry.createReader();
+            const entries = await new Promise(resolve => {
+                entryReader.readEntries(entries => resolve(entries));
+            });
+            await Promise.all(entries.map(entry => scanFiles(entry, tmpObject)));
+            break;
+        case (entry.isFile) :
+            tmpObject.push(entry);
+            break;
+    }
+}
+
+holder.addEventListener("dragover", event => {
+        event.preventDefault();
+    }, false);
+
+holder.addEventListener("drop", async event => {
+        //console.dir(e);;
+        event.preventDefault();
+        const items = event.dataTransfer.items;
+        const results = [];
+        const promise = [];
+        for (const item of items) {
+            const entry = item.webkitGetAsEntry();
+            promise.push(scanFiles(entry, results));
+        }
+        await Promise.all(promise);
+        console.log(results); //テスト表示
+
+    }, false);
+  </script>
+</body>
+</html>
